split code to improve tree shaking and resolve webpack error
something about pino.transport not being a function
