+++ src/components/mods/modCarousel.tsx
@@ -1,5 +1,6 @@
 import { Carousel } from "@mantine/carousel";
-import { type ModImageUrls, useModImageUrls } from "~/hooks/globalContexts/modImageUrls";
+import { useModImageUrls } from "~/hooks/globalContexts/modImageUrls/useModImageUrls";
+import type { ModImageUrls } from "~/hooks/globalContexts/modImageUrls/constsAndTypes";
 import { createStyles } from "@mantine/core";
 import { Image } from "@mantine/core";      //TODO!: replace with nextjs Image component once next.config.mjs is fixed
 // import Image from "next/image";
