+++ src/hooks/globalContexts/modDownloadUrl/constAndTypes.ts
@@ -1,0 +1,7 @@
+
+export type NewestFileId = string;
+
+
+
+
+export const GAMEBANANA_MOD_DOWNLOAD_BASE_URL = "everest:https://gamebanana.com/mmdl/";
