+++ src/hooks/globalContexts/modImageUrls/constsAndTypes.ts
@@ -1,0 +1,7 @@
+
+export type ModImageUrls = string[];
+
+
+
+
+export const GAMEBANANA_MOD_IMAGES_BASE_URL = "https://images.gamebanana.com";
